Add some testing stuff (13.)
While it's required that there are tests, it is obviously good
to still have them, and it's very much nicer to have some
tests that can be automatically run rather than having
to manually POST and GET and DELETE and so on for testing.
Didn't add all the wanted tests just yet, rather commit now
when it's in an okay state.