Add error handler middleware (13.)
Also a test to ensure correct statuses are returned.

Fantastic that sequelize apparently just attempts to parse
whatever's passed as a string if the type is set to text,
so setting title to 100 (as a number) is completely fine.
It also does this BEFORE validation, so tough. Could
validate before passing to sequelize, I guess.