Add partial "Add support for users to the application" (13.8)
Bit of a problem with the tests, where the User table is intermittenly
filled with something between a destroy and a create in the last
beforeEach, causing the first test ("Changes username") to fail because
the .create cannot create with a non-unique username. Not entirely
sure what is causing this, though was thinking that it could
be due to the Promise.alls, but then those should actually
wait for all the created asyncs to finish when the tests complete
succesfully, and that is what they're doing. Either way, it works
for now with a workaround, and the otherwise failing test doesn't
seem to be the actual problem.