Add partial "User linked to added blog", no test for whether userid is added (13.10)
The id doesn't seem to be added, not sure why that is just yet. In any case,
need to still add a test for that and implement it properly, I guess.
I'm not actually even putting the id in the all the test creates
either, obviously need to make the reference required.