Add "User linked to added blog" (13.10)
No surprise that the userId didn't work, as the key is actually
based on the referenced model's name (which was capitalised "User",
not "user"). Not very clear from the documentation, or at least
not on the "Associations" page. Even on the "Model Basics"
page, they literally use the capitalised "User" as the example name
for the model, which again should apparently result in a "UserId"
foreign key, not the "userId" that would be expected based on
the "Associations" docs.