Fix sending null/undefined token, calling error handlers (13.)
Figured out an issue and serendipitously found another, or two, in fact.
Firstly, the error handlers' if-else blocks where such that the
next(error) might not be reached, which caused the server to hang
when given a certain error. With this, I realised that null and undefined,
which were being sent as tokens to the POST blogs, were being replaced
by the actual token, which required a fix. In realising this, I also
noticed that the test that I had was not failing despite the
the errors, which was due to using the allSettled instead of all of Promise,
which does not throw automatically. I still wanted to use the allSettled,
as it allowed all the Promises to settle first before ending the test
(should use it elsewhere too, but haven't yet made the change),
so now it just finds all the promises that rejected and creates a
combined error message of those. Much nicer than having just the one
error message from whichever Promise rejects first.