Remove passwordHash from returned user data (13.)
The idea in general for everything returned by the API is that
it's not supposed to necessarily be public, or at least that's how
I consider it. Even then, there's not really a lot of sense in
returning the passwordHash as part of the data, as it's something
that would really only be used "in" the backend, so to speak, rather
than exposed by it. Sending it in a request also means that whoever
in the frontend is responsible for getting rid of it so as not to
expose it to an end user. Even further, it's still possible that the API
could be allowed to be used by end user developers at some point. I guess
it might ultimately be best to potentially develop with the assumption
that the API would in fact be used by some third party developer, but
it does depend.