Add "Make a migration that initializes the database" (13.17)
Hate it, absolutely hate it. Where's the DRY? Why am I having to
define everything twice? The migrationing part I understand,
the implementation of it is awful.

Well, okay, I did actually try and do it more neatly by using the
same attributes, and it does actually work, I think, though still
not amazingly well. The only
problem is that you can't pass "underscored" to the createTable,
so I would have to use the pascalCase or whatever format. This is fine,
really, as I shouldn't really even have to use Postgres directly, but
basically it was/is just a mess to work with, trying to figure out
whether something's in snake case or pascal case etc. Could certainly make the
change back, but it's working now and I cannot be bothered at the
moment, too much headache for one dev session.