Add "Now add functionality to the application to support the reading list" (13.20)
Not sure whether I'm supposed to use the :username or :id. The
assignment asks to "modify" the :id route, but this has not until
now existed. Obviously the routers make no distinction between
what the actual param is called, but I do, and the username is not the
same as the id. In any case, I'm going based on the specification,
though of course I would otherwise clarify if possible.

Really need to fix up some of the stuff. Main thing would be to finally
change the model names to camelCase, as I don't have them the way
they were intended to be and that's causing some annoyance.