Add partial "Server-side sessions", user disabled state (13.24)
Decided to add it as suggested, but now it doesn't seem very sensible
as I don't see how I'm going to be including it. The main aim of this
exercise is to add more robust sessions, but the validity of a session
is still, as far as I'm able to think, going to be checked using
a token. If there isn't a token associated with a user or such a token
is invalid for whatever reason, the user is already considered "disabled",
and if there is a token and it's valid, then the user is not disabled.
This does not require a separate disabled state for a user.

Of course, the disabled state could be used for other things,
presumably for admins being able to ban a user, for example. This is not
part of the specification for this exercise, however, so is not relevant.
For that matter, it would be more practical to have a more general
state, some status for a user, such as "disabled", "user", "admin",
or "superadmin".

On another note, how to allow the frontend to save the state of a user?
Technically the only thing needed would be an id of some sort that can
be matched with the sessions table, but in practice this would be
highly vulnerable to attacks. In practice, a JWT can still be used,
and in fact no changes are necessarily needed to the current setup.
By having the token contain the username of the user, the sessions
table can contain the username and the session can then be found using
the passed username. However, user ID (in the JWT token) is a more
practical value to use, allowing for neater connections.