Add partial "Server-side sessions", sessions table (13.24)
Not yet functionally included apart from having a session be
added during login.

Things required:
- Use bearer token to check user session validity from table
    - Current interface should not require many changes, only
    changes in the routes.
- Ensure that only usernames in the sessions table with still valid
sessions are allowed to perform restricted actions
    - Add methods to session class for checking validity / invalidating
    state

I'm currently not planning on actually implementing a automatic limited
validity period for the sessions, as this was not the case previously
either. It should be easy enough to add such an automatic setup
later on, though. As mentioned, the main use of the server-side sessions
is that access can be revoked regardless of expiration, which is not
the case if client-side tokens with no expiration are used.