Add partial "Server-side sessions", allow multiple sessions for one user (13.24)
I realised that having just the one session is going to be very difficult.
One issue is that logging out anywhere would mean that a logout would
happen everywhere for that user. Not only is this annoying even as is,
but someone getting hold of a valid token of that user could continue
to log out. Even worse, they could actually access the user's content
using that token if the user logs in somewhere else. Basically,
major security issue. Luckily I realised it fairly quickly,
and luckily I'm not actually developing a used product.

Due to just using the one session originally, the idea would have been
to just maintain each session row, but invalidate and validate it
when necessary. Now of course the idea would be to probably just
delete any invalid (out of date) sessions, and at least during logout a session
should be deleted to prevent a token being used further. Using invalidation
as a mechanism outside just testing is not good, because
again it allows a token to still be potentially used. However, there
is no route presently that would allow these operations anyway.